+++ projects/swagger-schematics/helpers/config.ts
@@ -6,7 +6,6 @@ type swaggerSchemaKeys = keyof SwaggerApiSchema;
 
 /** Default configuration values */
 const DEFAULT_CONFIG: Partial<SwaggerApiSchema> = {
-    framework: 'angular',
     scopeEndpointsWithTags: false,
 };